+++ CHANGELOG.md
@@ -11,7 +11,8 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Changed
 
-- Provider now throws an exception if location was not found.
+- Provider now thrown exception if location was not found.
+- Use PSR-17 HTTP factories instead of HTTPlug internal versions
 
 ## 0.2.0 - 2017-04-16