+++ CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Changed
 
-- Provider now thrown exception if location was not found.
+- Provider now throws exception if location was not found.
 - Use PSR-17 HTTP factories instead of HTTPlug internal versions
 
 ## 0.2.0 - 2017-04-16