+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
-## 0.3.0 - unreleased
+## 0.3.0 - 2017-05-31
+
+### Added
+
+- Support for Google browserlocation API
 
 ### Changed
 
-- Provider now thrown exception if location was not found.
+- Provider now throws exception if location was not found.
 - Use PSR-17 HTTP factories instead of HTTPlug internal versions
 
 ## 0.2.0 - 2017-04-16